KS-11-2 has been completed, please approve the pull request.
