+++ src/main/kotlin/lesson11/lesson11_task2.kt
@@ -13,3 +13,31 @@ fun main() {
     user1.writeBio()
     user1.changePassword()
 }
+
+class User2(
+    val id: Int,
+    val login: String,
+    var password: String,
+    val email: String,
+    var bio: String? = null
+) {
+    fun printUserData() {
+        println("Найден пользователь $login с почтой email $email")
+    }
+
+    fun writeBio() {
+        bio = readln()
+    }
+
+    fun changePassword() {
+        println("Для изменения пароля введите старый пароль:")
+        val inputPassword = readln()
+        if (inputPassword == password) {
+            println("Введите новый пароль")
+            password = readln()
+            println("Новый пароль успешно сохранен")
+        } else {
+            println("Старый пароль неверен")
+        }
+    }
+}
